requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-dotenv==1.4.2
djangorestframework==3.10.1
gunicorn==19.9.0
psycopg2-binary==2.8.3
pytz==2019.1
sqlparse==0.3.0
